Implement `Area` enums, structs, and classes
Perhaps these names should be prepended with `Draw` like the font
types, or perhaps the font types should drop the `Draw` prefix. One of
the two should probably happen for the sake of consistency. It’s not
documented in `libui`, so I’m not sure if all the `uiDrawArea…`
functions apply to any area, or if some are specific to scrolling
areas. I figure `uiAreaScrollTo` would be scrolling specific, but I’m
not sure about `uiAreaSetSize`.