Don't destroy a control when the class deinitializes
When the class goes out of scope it will be destroyed, but if the
control has been added to a parent, then the C control will still be
alive and active. In general, we can’t say that the Swift class has the
same lifetime as the C object, so don’t enforce that.