AppKit is only available on macOS
On Linux we need GTK, and on maxOS we need AppKit. AppKit is actually a
dependency for `clibui` and GTK is included therein, but the Swift
package manager doesn’t currently support adding explicit dependencies
on builtin swift packages. To get it linked in, then, I have to add
`import AppKit` to a Swift source. `clibui` doesn’t have any Swift
sources, so I have to add it here, which is really unfortunate.