Note the specific version of Swift 3
This package will not compile with Preview 5 and below. Between preview 5 and 6 they removed `UnsafeMutablePointer<Void>` and changed it to `UnsafeMutableRawPointer`, and they added `@escaping` and required that I explicitly state it on the event handlers.

Once Swift 3 is officially released I will remove this caveat and assume no one is using previews or GM candidates or whatever else they call the pre-release versions.